Revert "Revert "Improve IFile.getLineSeparator() API""
This reverts commit 166a3554857885f0ae81cff2f739847e08c9b1fa.
The previous commit was missing explanation about how to revert.

This restored changes restore the advantages of API improvement:
* Consistency with similar methods (ie IFile.getCharset)
* Easier OO navigation and discovery, as users are more likely to look up
state info for a given object by dereferencing it. So file... can be completed
easily to the expected getLineSeparator, while a developer wouldn't naturally
guess which utility class can provide the feature
* Better isolation of concerns: the lineSeparator is a state info of the IFile
or IProject, so let's keep related code in their implementation
